+++ src/pages/GroupEdit.vue
@@ -40,9 +40,14 @@
         </v-stepper-items>
       </v-stepper>
 
-      <DeleteConfirmationDialog v-model="dialog" @confirm="deleteGroup">
-        Delete this group
-      </DeleteConfirmationDialog>
+      <v-card class="px-5 pt-5 pb-3">
+        <span class="headline mr-3">
+          Delete this group
+        </span>
+        <v-btn small depressed color="error" class="mb-2" @click="deleteGroup">
+          Delete
+        </v-btn>
+      </v-card>
     </template>
   </v-container>
 </template>
@@ -57,7 +62,6 @@ import FormBackButton from '@/components/shared/FormBackButton.vue'
 import ProgressCircular from '@/components/shared/ProgressCircular.vue'
 import LoadFailedText from '@/components/shared/LoadFailedText.vue'
 import RepositoryFactory from '@/repositories/RepositoryFactory'
-import DeleteConfirmationDialog from '@/components/shared/DeleteConfirmationDialog.vue'
 
 const GroupsRepo = RepositoryFactory.get('groups')
 
@@ -69,13 +73,11 @@ const GroupsRepo = RepositoryFactory.get('groups')
     FormBackButton,
     ProgressCircular,
     LoadFailedText,
-    DeleteConfirmationDialog,
   },
 })
 export default class GroupEdit extends Vue {
   status: 'loading' | 'loaded' | 'error' = 'loading'
   canEdit = true
-  dialog = false
   valid = false
   step = 1
 
@@ -114,6 +116,10 @@ export default class GroupEdit extends Vue {
   }
 
   async deleteGroup() {
+    const confirmed = window.confirm(
+      'この操作は取り消せません。本当にこのグループを削除してもよろしいですか？'
+    )
+    if (!confirmed) return
     const groupId = this.$route.params.id
     try {
       await GroupsRepo.$groupId(groupId).delete()
